# src/main/java/com/employee/entity/response/TeamData.java
package com.employee.entity.response;

import com.fasterxml.jackson.annotation.JsonInclude;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.util.List;

@Builder
@Data
@AllArgsConstructor
@NoArgsConstructor
public class TeamData {
    private String managerId;
    private String name;
    private List<String> members;
}